Mark methods as invoked on a task in ReachabilityAnalysisEngine.addRootMethod.
